Renaming useResolver
