+++ Composer/packages/client/__tests__/components/TestController/publish-luis-modal.test.tsx
@@ -7,37 +7,36 @@ import { act } from '@bfc/test-utils/lib/hooks';
 
 import { PublishLuis } from '../../../src/pages/language-understanding/publish-luis-modal';
 import { projectIdState, botNameState, settingsState, dispatcherState } from '../../../src/recoilModel';
+import { renderWithRecoil } from '../../testUtils';
 
 describe('<PublishLuis />', () => {
   it('should render the <PublishLuis />', async () => {
     const onDismiss = jest.fn(() => {});
     const onPublish = jest.fn(() => {});
     const setSettingsMock = jest.fn(() => {});
+    const recoilInitState = ({ set }) => {
+      set(dispatcherState, {
+        setSettings: setSettingsMock,
+      });
+      set(projectIdState, '12345');
+      set(botNameState, 'sampleBot0');
+      set(settingsState, {
+        luis: {
+          name: '',
+          authoringKey: '12345',
+          authoringEndpoint: 'testAuthoringEndpoint',
+          endpointKey: '12345',
+          endpoint: 'testEndpoint',
+          authoringRegion: 'westus',
+          defaultLanguage: 'en-us',
+          environment: 'composer',
+        },
+      });
+    };
 
-    const { getByText } = render(
-      <RecoilRoot
-        initializeState={({ set }) => {
-          set(dispatcherState, {
-            setSettings: setSettingsMock,
-          });
-          set(projectIdState, '12345');
-          set(botNameState, 'sampleBot0');
-          set(settingsState, {
-            luis: {
-              name: '',
-              authoringKey: '12345',
-              authoringEndpoint: 'testAuthoringEndpoint',
-              endpointKey: '12345',
-              endpoint: 'testEndpoint',
-              authoringRegion: 'westus',
-              defaultLanguage: 'en-us',
-              environment: 'composer',
-            },
-          });
-        }}
-      >
-        <PublishLuis workState={0} onDismiss={onDismiss} onPublish={onPublish} />
-      </RecoilRoot>
+    const { getByText } = renderWithRecoil(
+      <PublishLuis workState={0} onDismiss={onDismiss} onPublish={onPublish} />,
+      recoilInitState
     );
 
     expect(getByText('What is the name of your bot?')).not.toBeNull();
